Importing source code
I've added the source for what I have currently. The problems with the
parser as it stands now is that it does not get some recipes if they are
on a different page. For example, if you have to click the next button
to see the recipes on FFXIAH.com they will not be parsed.